nat: fix dhcp client on outside interface with output feature
There was an attempt to fix this problem in the commit:
229d8e7630bcc6c67aefe00dda09ed23d2cc8c9a

But checking the LOCALLY_ORIGINATED flag didn't work because this flag
gets reset before it can reach the NAT nodes.

With this commit, replace the check for the LOCALLY_ORIGINATED flag
with a check to see if the packet is a DHCP broadcast.

Type: fix

Change-Id: I069c08a785b5988b10192f528e4f9c4c7cc2f8a3
Signed-off-by: Alexander Chernavin <achernavin@netgate.com>